style: format code

+++ src/LendingPoolProxy.sol
@@ -96,8 +96,7 @@ contract LendingPoolProxy {
     mapping(address => Loan[]) public userLoans;
 
     // mapping to tract liquidation history
-    mapping(address => mapping(uint256 => LiquidationEvent))
-        public liqudationHistory;
+    mapping(address => mapping(uint256 => LiquidationEvent)) public liqudationHistory;
 
     // mapping to track interest ratte model per asset
     mapping(address => InterestCalculator) public interestModels;
@@ -136,22 +135,11 @@ contract LendingPoolProxy {
         _;
     }
 
-    event CollateralDeposited(
-        address indexed _sender,
-        address indexed _token,
-        uint256 _value
-    );
-    event Borrowed(
-        address indexed _sender,
-        address indexed _token,
-        uint256 _value
-    );
+    event CollateralDeposited(address indexed _sender, address indexed _token, uint256 _value);
+    event Borrowed(address indexed _sender, address indexed _token, uint256 _value);
 
     event LoanRepaid(
-        address indexed _sender,
-        address indexed _token,
-        uint256 _effectivePayment,
-        int256 _interestPayment
+        address indexed _sender, address indexed _token, uint256 _effectivePayment, int256 _interestPayment
     );
 
     constructor() {
@@ -233,11 +221,7 @@ contract LendingPoolProxy {
         }
 
         bool tokenExist = false;
-        for (
-            uint256 i = 0;
-            i < userAccount[msg.sender].collateralAssets.length;
-            i++
-        ) {
+        for (uint256 i = 0; i < userAccount[msg.sender].collateralAssets.length; i++) {
             if (token == userAccount[msg.sender].collateralAssets[i]) {
                 tokenExist = true;
                 break;
@@ -254,10 +238,7 @@ contract LendingPoolProxy {
         emit CollateralDeposited(msg.sender, msg.value);
     }
 
-    function borrow(
-        address token,
-        uint256 amount
-    ) external sufficientCollateral(msg.sender) {
+    function borrow(address token, uint256 amount) external sufficientCollateral(msg.sender) {
         if (msg.value <= 0 || amount <= 0) {
             revert LendingPool__InvalidDepositAmount();
         }
@@ -271,11 +252,7 @@ contract LendingPoolProxy {
         }
 
         bool tokenExists = false;
-        for (
-            uint i = 0;
-            i < userAccounts[msg.sender].borrowedAssets.length;
-            i++
-        ) {
+        for (uint256 i = 0; i < userAccounts[msg.sender].borrowedAssets.length; i++) {
             if (userAccounts[msg.sender].borrowedAssets[i] == token) {
                 tokenExists = true;
                 break;
@@ -289,10 +266,7 @@ contract LendingPoolProxy {
         userBorrows[msg.sender][token] += amount;
         borrowTimestamp[msg.sender][token] = block.timestamp;
         totalBorrows[token] += amount;
-        userAccounts[msg.sender].totalBorrowsUSD += getAssetPrice(
-            amount,
-            AggregatorV3Interface(token)
-        );
+        userAccounts[msg.sender].totalBorrowsUSD += getAssetPrice(amount, AggregatorV3Interface(token));
 
         Loan memory userLoan = Loan({
             borrowAsset: token,
@@ -335,34 +309,23 @@ contract LendingPoolProxy {
         uint256 totalDebt = userBorrows[msg.sender][token] + interest;
 
         uint256 effectiveRepayment = amount > totalDebt ? totalDebt : amount;
-        uint256 principalpayment = effectiveRepayment > interest
-            ? effectiveRepayment - interest
-            : 0;
-        uint256 interestRepayment = effectiveRepayment > interest
-            ? interest
-            : effectiveRepayment;
+        uint256 principalpayment = effectiveRepayment > interest ? effectiveRepayment - interest : 0;
+        uint256 interestRepayment = effectiveRepayment > interest ? interest : effectiveRepayment;
 
         totalBorrows[token] -= principalpayment;
         userBorrows[msg.sender][token] -= principalRepayment;
 
         uint256 remainingRepayment = principalRepayment;
-        for (
-            uint i = 0;
-            i < userLoans[msg.sender].length && remainingRepayment > 0;
-            i++
-        ) {
+        for (uint256 i = 0; i < userLoans[msg.sender].length && remainingRepayment > 0; i++) {
             Loan storage loan = userLoans[msg.sender][i];
 
             if (loan.active && loan.borrowAsset == token) {
-                uint256 loanRemainingDebt = loan.borrowAmount -
-                    loan.amountRepaid;
+                uint256 loanRemainingDebt = loan.borrowAmount - loan.amountRepaid;
 
                 if (loanRemainingDebt > 0) {
                     // Calculate how much to apply to this loan
-                    uint256 loanRepayment = remainingRepayment >
-                        loanRemainingDebt
-                        ? loanRemainingDebt
-                        : remainingRepayment;
+                    uint256 loanRepayment =
+                        remainingRepayment > loanRemainingDebt ? loanRemainingDebt : remainingRepayment;
 
                     // Update loan record
                     loan.amountRepaid += loanRepayment;
@@ -380,37 +343,21 @@ contract LendingPoolProxy {
             _removeAssetFromBorrowedAssets(msg.sender, token);
         }
 
-        emit LoanRepaid(
-            msg.sender,
-            token,
-            effectiveRepayment,
-            interestRepayment
-        );
+        emit LoanRepaid(msg.sender, token, effectiveRepayment, interestRepayment);
     }
 
-    function calculatePositionHealth(
-        address user
-    ) public view returns (PositionHealth memory) {
+    function calculatePositionHealth(address user) public view returns (PositionHealth memory) {
         PositionHealth memory health;
 
         // calculate collateral value
         uint256 collateralValueUSD = 0;
-        for (
-            uint256 i = 0;
-            i < userAccounts[user].collateralAssets.length;
-            i++
-        ) {
+        for (uint256 i = 0; i < userAccounts[user].collateralAssets.length; i++) {
             address asset = userAccounts[user].collateralAssets[i];
             uint256 amount = userCollateral[user][asset];
-            uint256 valueUSD = getAssetPrice(
-                amount,
-                AggregatorV3Interface(asset)
-            );
+            uint256 valueUSD = getAssetPrice(amount, AggregatorV3Interface(asset));
 
             // apply collateral factor
-            valueUSD =
-                (valueUSD * assetConfigs[assert].collateralFactor) /
-                1000;
+            valueUSD = (valueUSD * assetConfigs[assert].collateralFactor) / 1000;
 
             collateralValueUSD += valueUSD;
         }
@@ -425,10 +372,7 @@ contract LendingPoolProxy {
             uint256 interest = calculateAccuredIntest(user, asset, amount);
 
             uint256 totalOwed = interest + amount;
-            uint256 valueUSD = getAssetPrice(
-                amount,
-                AggregatorV3Interface(asset)
-            );
+            uint256 valueUSD = getAssetPrice(amount, AggregatorV3Interface(asset));
             borrowValueUSD += valueUSD;
         }
 
@@ -446,11 +390,7 @@ contract LendingPoolProxy {
         return health;
     }
 
-    function calculateAccuredInterest(
-        address user,
-        address asset,
-        uint256 amount
-    ) public view returns (uint256) {
+    function calculateAccuredInterest(address user, address asset, uint256 amount) public view returns (uint256) {
         InterestCalculator memory assetInterest = interestModels[asset];
         uint256 borrowedAmount = userBorrows[user][asset];
 
@@ -474,10 +414,7 @@ contract LendingPoolProxy {
         uint256 utilizationRate = calculateUtilizationRate(asset);
 
         // Calculate the applicable interest rate based on utilization
-        uint256 interestRate = calculateApplicableInterestRate(
-            utilizationRate,
-            interestModel
-        );
+        uint256 interestRate = calculateApplicableInterestRate(utilizationRate, interestModel);
 
         // Convert annual interest rate to per-second rate
         // Division by 10000 to convert from basis points to decimal
@@ -486,15 +423,12 @@ contract LendingPoolProxy {
 
         // Calculate compound interest: principal * ((1 + ratePerSecond) ^ elapsedTime) - principal
         // For Solidity, we use the approximation: principal * (1 + ratePerSecond * elapsedTime)
-        uint256 interest = (borrowedAmount * ratePerSecond * elapsedTime) /
-            1e10;
+        uint256 interest = (borrowedAmount * ratePerSecond * elapsedTime) / 1e10;
 
         return interest;
     }
 
-    function calculateUtilizationRate(
-        address asset
-    ) internal view returns (uint256) {
+    function calculateUtilizationRate(address asset) internal view returns (uint256) {
         uint256 totalDepositForAsset = totalDepoits[asset];
 
         if (totalDepositForAsset == 0) {
@@ -508,36 +442,28 @@ contract LendingPoolProxy {
         return (totalBorrowsForAsset * 10000) / totalDepositsForAsset;
     }
 
-    function calculateApplicabeInterestRate(
-        uint256 utilizationRate,
-        InterestCalculator memory model
-    ) internal pure returns (uin256) {
+    function calculateApplicabeInterestRate(uint256 utilizationRate, InterestCalculator memory model)
+        internal
+        pure
+        returns (uin256)
+    {
         if (utilizationRate <= model.optimalUtilizationRate) {
             // Calculate: baseRate + (utilizationRate * slope1) / optimalUtilizationRate
-            return
-                model.baseRatePerYear +
-                ((utilizationRate * model.slopeRate1) /
-                    model.optimalUtilizationRate);
+            return model.baseRatePerYear + ((utilizationRate * model.slopeRate1) / model.optimalUtilizationRate);
         }
         // If utilization is above the optimal rate, use the second (steeper) slope
         else {
             // Calculate: baseRate + slope1 + ((utilizationRate - optimalRate) * slope2) / (10000 - optimalRate)
-            uint256 excessUtilization = utilizationRate -
-                model.optimalUtilizationRate;
+            uint256 excessUtilization = utilizationRate - model.optimalUtilizationRate;
             uint256 remainingUtilization = 10000 - model.optimalUtilizationRate;
 
-            return
-                model.baseRatePerYear +
-                model.slopeRate1 +
-                ((excessUtilization * model.slopeRate2) / remainingUtilization);
+            return model.baseRatePerYear + model.slopeRate1
+                + ((excessUtilization * model.slopeRate2) / remainingUtilization);
         }
     }
 
-    function getAssetPrice(
-        uint256 amount,
-        AggregatorV3Interface asset
-    ) public view returns (uint256) {
-        (, int256 price, , uint256 updatedAt, ) = asset.latestRoundData();
+    function getAssetPrice(uint256 amount, AggregatorV3Interface asset) public view returns (uint256) {
+        (, int256 price,, uint256 updatedAt,) = asset.latestRoundData();
 
         if (price <= 0) {
             revert LendingPool__ErrorInPrice();
@@ -547,24 +473,18 @@ contract LendingPoolProxy {
             revert LendingPool__StalePriceData();
         }
 
-        uint256 assetInUSD = (amount * uint256(price)) /
-            (10 ** asset.decimals());
+        uint256 assetInUSD = (amount * uint256(price)) / (10 ** asset.decimals());
 
         return assetInUSD;
     }
 
-    function _removeAssetFromBorrowedAssets(
-        address user,
-        address token
-    ) internal {
+    function _removeAssetFromBorrowedAssets(address user, address token) internal {
         UserAccount storage account = userAccounts[user];
-        for (uint i = 0; i < account.borrowedAssets.length; i++) {
+        for (uint256 i = 0; i < account.borrowedAssets.length; i++) {
             if (account.borrowedAssets[i] == token) {
                 // Replace with the last element and pop
                 if (i < account.borrowedAssets.length - 1) {
-                    account.borrowedAssets[i] = account.borrowedAssets[
-                        account.borrowedAssets.length - 1
-                    ];
+                    account.borrowedAssets[i] = account.borrowedAssets[account.borrowedAssets.length - 1];
                 }
                 account.borrowedAssets.pop();
                 break;
